+++ yabause/src/tsunami/Tsunami.c
@@ -1,9 +1,9 @@
-/* 
+/*
    Tsunami.c
 */
-/* 
+/*
    The MIT License (MIT)
-   
+
    Copyright (c) 2014 James A. McCombe
 
    Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -12,17 +12,17 @@
    to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
    copies of the Software, and to permit persons to whom the Software is
    furnished to do so, subject to the following conditions:
-   
+
    The above copyright notice and this permission notice shall be included in all
    copies or substantial portions of the Software.
-   
+
    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-   SOFTWARE.   
+   SOFTWARE.
 */
 
 #include <stdlib.h>
@@ -47,7 +47,7 @@
 /* ================================================================================ */
 /* Tsunami Internal Logging */
 /* Tsunami maintains a compact internal log of the value changes over time in
-   memory.  It stores this in a circular buffer of a user-defined size.  This 
+   memory.  It stores this in a circular buffer of a user-defined size.  This
    allows for a fixed sliding window of time to be captured.  Once the application
    finishes the timeline, it will dump this internal buffer to a VCD file on disk
    so that the data can be viewed using external tools.                             */
@@ -58,7 +58,7 @@ typedef enum _TsunamiLogEntryType {
 
 typedef struct _TsunamiLogEntry {
 	TsunamiLogEntryType  type;
-	
+
 	union {
 		struct {
 			uint64_t         time;
@@ -76,7 +76,7 @@ typedef struct _TsunamiLogEntry {
 /* ================================================================================ */
 /* Structure - A Tsunami "timeline" context.                                        */
 struct _TsunamiTimeline {
-	
+
 	char *name;        /* Name of the timeline */
 	char *filename;    /* VCD filename to write the timeline to on Finish() */
 
@@ -87,13 +87,13 @@ struct _TsunamiTimeline {
 		uint32_t         entry_last_index;  /* End of the circular buffer      */
 		uint32_t         entry_count;       /* Number of active entries        */
 	} log;
-	
+
 	TsunamiVariable var_root;  /* Variables - This is a hierarchy of nodes */
 
 	uint64_t        starting_usec; /* Microseconds at timeline start */
 	uint64_t        clk;           /* Current time slice             */
 	uint32_t        next_uid;      /* Variable unique ID allocator   */
-	
+
 	struct _TsunamiTimeline *next; /* Next timeline */
 };
 
@@ -120,8 +120,8 @@ static uint32_t _tsunami_context_is_initialised = 0;
 /* ================================================================================ */
 TsunamiLogEntry *TsunamiGetLogWritePtrAndAdvance(TsunamiTimeline *timeline)
 {
-	TsunamiLogEntry *entry = (timeline->log.entry + timeline->log.entry_last_index);	
-	
+	TsunamiLogEntry *entry = (timeline->log.entry + timeline->log.entry_last_index);
+
 	timeline->log.entry_last_index = ((timeline->log.entry_last_index + 1) % timeline->log.entry_alloc_count);
 	if (timeline->log.entry_count == timeline->log.entry_alloc_count)
 		timeline->log.entry_start_index = ((timeline->log.entry_start_index + 1) % timeline->log.entry_alloc_count);
@@ -131,9 +131,9 @@ TsunamiLogEntry *TsunamiGetLogWritePtrAndAdvance(TsunamiTimeline *timeline)
 	return entry;
 }
 
-void TsunamiDumpSignals_Traverse(FILE            *output_vcd_file, 
-								 TsunamiTimeline *timeline, 
-								 TsunamiVariable *var, 
+void TsunamiDumpSignals_Traverse(FILE            *output_vcd_file,
+								 TsunamiTimeline *timeline,
+								 TsunamiVariable *var,
 								 uint32_t         depth)
 {
 	uint32_t i;
@@ -143,19 +143,19 @@ void TsunamiDumpSignals_Traverse(FILE            *output_vcd_file,
 			fprintf(output_vcd_file, "   ");
 
 		if (var->head) {
-			fprintf(output_vcd_file, "$scope module %s $end\n", var->node_name);			
+			fprintf(output_vcd_file, "$scope module %s $end\n", var->node_name);
 		} else {
 			fprintf(output_vcd_file, "$var wire %i %s %s [%d:0] $end\n", var->size, var->uid, var->node_name, var->size);
 		}
-		
+
 		if (var->head) {
 			TsunamiDumpSignals_Traverse(output_vcd_file,
-										timeline, 
-										var->head, 
+										timeline,
+										var->head,
 										(depth + 1));
 			for (i = 0; i < depth; i ++)
-				fprintf(output_vcd_file, "   ");	
-			fprintf(output_vcd_file, "$upscope $end\n");		
+				fprintf(output_vcd_file, "   ");
+			fprintf(output_vcd_file, "$upscope $end\n");
 		}
 
 		var = var->next;
@@ -163,40 +163,40 @@ void TsunamiDumpSignals_Traverse(FILE            *output_vcd_file,
 }
 
 void TsunamiDumpRanges_Traverse(FILE            *output_vcd_file,
-								TsunamiTimeline *timeline, 
-								TsunamiVariable *var, 
+								TsunamiTimeline *timeline,
+								TsunamiVariable *var,
 								uint32_t         depth)
 {
 	while (var) {
 		if ((!var->head) && (var->range)) {
 			uint32_t i;
-			fprintf(output_vcd_file, "#%llu\n", (unsigned long long) timeline->clk);
+			fprintf(output_vcd_file, "#%lu\n", (long unsigned int) timeline->clk);
 			fprintf(output_vcd_file, "b");
-			for (i = 0; i < 32; i ++) 
+			for (i = 0; i < 32; i ++)
 				fprintf(output_vcd_file, "%c", (int) ('0' + ((var->range >> (63 - i)) & 0x1)));
-			fprintf(output_vcd_file, " %s\n", var->uid);			
+			fprintf(output_vcd_file, " %s\n", var->uid);
 		} else {
-			TsunamiDumpRanges_Traverse(output_vcd_file, 
-									   timeline, 
-									   var->head, 
-									   (depth + 1));			
+			TsunamiDumpRanges_Traverse(output_vcd_file,
+									   timeline,
+									   var->head,
+									   (depth + 1));
 		}
-			
+
 		var = var->next;
 	}
 }
 
 TsunamiTimeline *TsunamiFindTimeline_Internal(const char *timeline_name)
 {
-	TsunamiContext  *ctx      = &_tsunami_context;	
+	TsunamiContext  *ctx      = &_tsunami_context;
 	TsunamiTimeline *timeline = ctx->timeline_head;
-	
+
 	while (timeline) {
 		if (!strcmp(timeline_name, timeline->name))
 			return timeline;
 		timeline = timeline->next;
 	}
-	
+
 	return NULL;
 }
 
@@ -252,7 +252,7 @@ void TsunamiTerminateSignalHandler(int signal)
 		TsunamiFlushTimeline(timeline->name);
 		timeline = timeline->next;
 	}
-	
+
 	exit(1);
 }
 
@@ -261,7 +261,7 @@ void TsunamiInitialise(void)
 	if (!_tsunami_context_is_initialised) {
 		TsunamiContext *ctx = &_tsunami_context;
 		memset(ctx, 0, sizeof(TsunamiContext));
-		
+
 		/* Create mutex */
 #if _WIN32
 
@@ -279,7 +279,7 @@ void TsunamiInitialise(void)
 		signal(SIGSEGV, &TsunamiTerminateSignalHandler);
 		signal(SIGTERM, &TsunamiTerminateSignalHandler);
 		signal(SIGINT,  &TsunamiTerminateSignalHandler);
-		
+
 		_tsunami_context_is_initialised = 1;
 
 		printf("TSUNAMI: Initialised and installed termination signal handlers\n");
@@ -289,11 +289,11 @@ void TsunamiInitialise(void)
 void TsunamiStartTimeline(const char *timeline_name, const char *filename, uint32_t log_size_bytes)
 {
 	TsunamiContext  *ctx      = &_tsunami_context;
-	
+
 	TsunamiLock_Internal(ctx);
 	{
 		TsunamiTimeline *timeline = TsunamiFindTimeline_Internal(timeline_name);
-		
+
 		if (timeline) {
 			printf("TSUNAMI: Found existing timeline for \"%s\", flushing its contents and restarting it from scratch\n",
 				   timeline_name);
@@ -325,7 +325,7 @@ void TsunamiStartTimeline(const char *timeline_name, const char *filename, uint3
 		/* Setup the timeline */
 		timeline->log.entry_alloc_count = (log_size_bytes / sizeof(TsunamiLogEntry));
 		timeline->log.entry             = ((TsunamiLogEntry *) calloc(sizeof(TsunamiLogEntry), timeline->log.entry_alloc_count));
-		
+
 		/* Put in a zero timestamp into the change log to start off */
 		{
 			TsunamiLogEntry *entry = TsunamiGetLogWritePtrAndAdvance(timeline);
@@ -358,8 +358,8 @@ void TsunamiFlushTimeline(const char *timeline_name)
 
 				/* Dump signal hierarchy */
 				TsunamiDumpSignals_Traverse(output_vcd_file,
-											timeline, 
-											timeline->var_root.head, 
+											timeline,
+											timeline->var_root.head,
 											0);
 
 				/* Dump change deltas */
@@ -369,7 +369,7 @@ void TsunamiFlushTimeline(const char *timeline_name)
 					uint32_t         found_first_time = 0;
 
 					for (i = 0, j = timeline->log.entry_start_index; i < timeline->log.entry_count; i ++) {
-			
+
 						entry = (timeline->log.entry + j);
 
 						/* The first entry in the file should be a time stamp so wait for it */
@@ -378,7 +378,7 @@ void TsunamiFlushTimeline(const char *timeline_name)
 
 						switch (entry->type) {
 						case TsunamiLogEntryType_NewTime:
-							fprintf(output_vcd_file, "#%llu\n", (unsigned long long) entry->new_time.time);
+							fprintf(output_vcd_file, "#%lu\n", (long unsigned int) entry->new_time.time);
 							found_first_time = 1;
 							break;
 						case TsunamiLogEntryType_ChangeValue:
@@ -388,16 +388,16 @@ void TsunamiFlushTimeline(const char *timeline_name)
 							fprintf(output_vcd_file, " %s\n", entry->change_value.var->uid);
 							break;
 						}
-			
+
 					skip:
 						j = ((j + 1) % timeline->log.entry_alloc_count);
 					}
 				}
-		
+
 				/* Dump ranges (for signals which have ranges set) */
 				TsunamiDumpRanges_Traverse(output_vcd_file,
-										   timeline, 
-										   timeline->var_root.head, 
+										   timeline,
+										   timeline->var_root.head,
 										   0);
 
 				fclose(output_vcd_file);
@@ -416,7 +416,7 @@ void TsunamiFlushTimeline(const char *timeline_name)
 void TsunamiAdvanceTimeline_Internal(TsunamiTimeline *timeline)
 {
 
-	TsunamiLogEntry *entry    = TsunamiGetLogWritePtrAndAdvance(timeline);	
+	TsunamiLogEntry *entry    = TsunamiGetLogWritePtrAndAdvance(timeline);
 
 	/* Write a new time token into the log */
 	entry->type          = TsunamiLogEntryType_NewTime;
@@ -425,9 +425,9 @@ void TsunamiAdvanceTimeline_Internal(TsunamiTimeline *timeline)
 	/* Reset pulse variables */
 	{
 		TsunamiVariable *var = timeline->var_root.list_next;
-		
+
 		while (var) {
-			if (var->is_pulse) { 
+			if (var->is_pulse) {
 				TsunamiSetValue_Internal(timeline, var, 0);
 				var->last_value = 0;
 			}
@@ -482,7 +482,7 @@ TsunamiVariable *TsunamiFindVariable_Traverse_Internal(TsunamiTimeline *timeline
 			break;
 	}
 	len = i;
-	
+
 	/* Search this level for a matching scope name */
 	var = parent_var->head;
 	while (var) {
@@ -501,11 +501,11 @@ TsunamiVariable *TsunamiFindVariable_Traverse_Internal(TsunamiTimeline *timeline
 	/* Descend another level if needed */
 	if (var->head)
 		return TsunamiFindVariable_Traverse_Internal(timeline, (name + len + 1), (name_len - len - 1), var);
-	
+
 	return var;
 }
 
-TsunamiVariable *TsunamiAddVariable_Internal(TsunamiTimeline *timeline, 
+TsunamiVariable *TsunamiAddVariable_Internal(TsunamiTimeline *timeline,
 											 const char *full_name, uint32_t full_name_len,
 											 const char *name,      uint32_t name_len,
 											 TsunamiVariable *parent_var)
@@ -517,7 +517,7 @@ TsunamiVariable *TsunamiAddVariable_Internal(TsunamiTimeline *timeline,
 	/* A NULL var suggests starting traversal at root */
 	if (!parent_var)
 		parent_var = &timeline->var_root;
-	
+
 	/* Extract the name of this scope */
 	node_name_len = name_len;
 	for (i = 0; i < node_name_len; i ++) {
@@ -527,7 +527,7 @@ TsunamiVariable *TsunamiAddVariable_Internal(TsunamiTimeline *timeline,
 	if (i == node_name_len)
 		is_leaf = 1;
 	node_name_len = i;
-	
+
 	/* Search this level for a matching scope name */
 	var = parent_var->head;
 	while (var) {
@@ -537,7 +537,7 @@ TsunamiVariable *TsunamiAddVariable_Internal(TsunamiTimeline *timeline,
 		}
 		var = var->next;
 	}
-	
+
 	/* If we failed to find a match at this level, create this scope and
 		continue descent */
 	if (!var) {
@@ -579,18 +579,18 @@ TsunamiVariable *TsunamiAddVariable_Internal(TsunamiTimeline *timeline,
 			var->list_next          = timeline->var_root.list_next;
 		timeline->var_root.list_next = var;
 	}
-	
+
 	/* Descend another level if needed */
 	if (!is_leaf)
 		return TsunamiAddVariable_Internal(timeline, full_name, full_name_len,
 										   (name + node_name_len + 1), (name_len - node_name_len - 1), var);
-	
+
 	return var;
 }
 
 TsunamiVariable *TsunamiFindVariable_Internal(TsunamiTimeline *timeline, const char *name)
 {
-	TsunamiVariable *var      = NULL;	
+	TsunamiVariable *var      = NULL;
 	uint32_t         name_len;
 
 	/* Find variable */
@@ -600,14 +600,14 @@ TsunamiVariable *TsunamiFindVariable_Internal(TsunamiTimeline *timeline, const c
 	/* Create a new variable, if existing one was not found */
 	if (!var)
 		var = TsunamiAddVariable_Internal(timeline, name, name_len, name, name_len, NULL);
-	
+
 	return var;
 }
 
 void TsunamiSetValue_Internal(TsunamiTimeline *timeline, TsunamiVariable *var, uint64_t value)
 {
-	/* Only write the variable if the data has changed for this signal */ 
-	if ((var->last_value != value) || (!var->is_defined)) {	
+	/* Only write the variable if the data has changed for this signal */
+	if ((var->last_value != value) || (!var->is_defined)) {
 
 		TsunamiLogEntry *entry = TsunamiGetLogWritePtrAndAdvance(timeline);
